+++ NEWS.md
@@ -8,7 +8,7 @@
 
 - Added new functions `Rscript` and `Rcmd` as wrappers of `system2('Rscript')` and `system2('R', 'CMD')`, respectively.
 
-- Added a function `install_dir()` to install a source package from a directory.
+- Added a new function `install_dir()` to install a source package from a directory.
 
 - Added a new function `file_string()` to read a text file (encoded in UTF-8) and return its content a single character string (lines concatenated by `\n`).